requirements.txt
Libraries Used

numpy
Matplotlib
hdbscan
kafka-python
scikit-learn